backend/app/rag/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardcap-rag"
version = "2.0.0"
description = "LangGraph-based Agentic RAG system for intelligent email PII auto-masking"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.3",
    "konlpy>=0.6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "rank-bm25>=0.2.2",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.0.20",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "jsonschema>=4.20.0",
    "colorama>=0.4.6",
    "rich>=13.7.0",
    "urllib3<2.0.0",
    "certifi",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[tool.setuptools]
packages = ["agent", "api", "scripts", "schemas"]

[tool.setuptools.package-data]
schemas = ["*.json"]
